more fixes
vehicles no longer use node teleporting when teleported by Teleporter Bullets as it didn't work well
unstuck code: fixed collision group not being set back correctly depending on ducking state
vphysics things teleported by Teleporter Bullets now use the "proper" Teleport() as opposed to SetAbsOrigin() which tends to have issues
fixed mounted guns acting weird when teleported by Teleporter Bullets
unstuck searching now has more angular granularity once it is testing spots that are more than half the distance to the max distance. (Teleporter Bullets, shoot the curve of the first pipe in d1_canals_12 with the airboat gun)

+++ sp/src/game/server/baseentity.h
@@ -1539,7 +1539,7 @@ public:
 public:
 	bool GetUnstuck(float flMaxDist, int flags = 0);
 	bool FindPassableSpace(const Vector direction, float step, Vector& oldorigin, CUtlVector<Vector> &vecBadDirections, int flags = 0);
-	bool FindOffsetSpot(Vector forward, int FFlip, Vector right, int RFlip, Vector up, int UFlip, Vector& vecGoodSpot, int flDist, CUtlVector<Vector> &vecBadDirections, int flags = 0);
+	bool FindOffsetSpot(Vector forward, float FFlip, Vector right, float RFlip, Vector up, float UFlip, Vector& vecGoodSpot, int flDist, CUtlVector<Vector> &vecBadDirections, int flags = 0);
 	bool CheckIfBelowGround(Vector vecPos, bool bNoDebug);
 	bool PutAtNearestNode(float flMaxDist, bool bNoDebug);
 	// Add a discontinuity to a step
